backend/requirements.txt
Flask==3.1.0
Flask-CORS==3.0.10
psycopg2-binary==2.9.8
Flask-SQLAlchemy==3.0.5
flask_socketio